IMPALA-11250: Adapt Webserver.TestWithSpnego for curl >= 7.64.0
Curl changed behaviors in 7.64.0. Previously, it would attempt
to connect without authentication, and for the SPNEGO test,
this would result in a failed attempt. Starting in Curl 7.64,
there is no attempt without authentication, so the count of
failed attempts is one less.

This detects whether the curl version is 7.64.0 or above
and adjusts the count of failed attempts accordingly.

Testing:
 - Ran Webserver.TestWithSpnego on Ubuntu 20 (curl version 7.68)
 - Ran Webserver.TestWithSpnego on Ubuntu 16 (curl version 7.47)

Change-Id: Icc9b92b334f878792ee649dafa30aed617e55589
Reviewed-on: http://gerrit.cloudera.org:8080/18832
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>