IMPALA-11226: Add script to simplify resolving minidumps
This adds the resolve_minidumps.py script to
simplify resolving minidumps under ideal circumstances.
This is designed to handle cases where the binary
and libraries are in identical locations to when
the minidump was created. This is true for developer
environments and at the end of Jenkins jobs.

This uses Breakpad's minidump_dump utility to get a
list of the binaries/libraries that the minidump
references. It uses that list to dump all the
symbols to a temporary directory. Then it uses
the symbols to resolve the minidump.

Since it is dumping symbols for all referenced
libraries, it resolves symbols to the maximum
extent possible.

This adds a step to bin/jenkins/finalize.sh to use
this new script to resolve minidumps. The old method
can be removed in a subsequent change.

Testing:
 - Ran locally on a minidump generated by sending
   SIGUSR1 to local impalad
 - Tested with a Centos 7 job using Python 3.6
   and verified the minidump output
 - Tested resolving a minidump from a binary with
   compressed debug info

Change-Id: I0f8fdcb8ca89d0904dc8ec69337e3d5dfdd54adf
Reviewed-on: http://gerrit.cloudera.org:8080/18918
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>