IMPALA-11548: Share same TTransport object for RPC input and output
After the CPP thrift upgrade to version 0.16.0 by IMPALA-11384, we are
seeing long-running connections closed by the Thrift server due to
MaxMessageSize reached. Additional tracing log in the thrift library and
TAcceptQueueServer reveals that TTransport's message size counter was
continuously decreasing over the lifetime of the connection and was
never reset. Further investigation points down to having separate input
and output TTransport object in TAcceptQueueServer::SetupConnection as
the root cause.

This patch fixes the issue by sharing the same TTransport object for
both Thrift RPC input and output in TAcceptQueueServer::SetupConnection.
Using the same TTransport causes the counter to be decremented and reset
on the same object. TSaslTransport had a caching algorithm in
TSaslServerTransport::Factory::getTransport() that guaranteed that the
underlying transport was shared between input and output. This is no
longer necessary and has been removed.

Testing:
- Pass core tests.

Change-Id: I199d6b0c6c62e940e131eb39d38a8a51b36e11c4
Reviewed-on: http://gerrit.cloudera.org:8080/18938
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>