IMPALA-11159: Fix AttributeError in TestAsyncDDLTiming.test_ctas
TestAsyncDDLTiming.test_ctas is likely to be flaky, additionally it
throws AttributeError when fails which hides the details of the failure.
This commit fixes the AttributeError which is caused by the Impyla tests,
the Impyla handler is a HiveServer2Cursor object that does not have an
'id' attribute. Query string can be printed instead of the query id.

Testing:
 - Triggered test failure manually and verified the result.

Change-Id: Ibe78b3a3ee070a1edeaef5cf7ae8d9565a10141d
Reviewed-on: http://gerrit.cloudera.org:8080/18933
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>