IMPALA-11599/IMPALA-11605: GCC 10: Fix gdb and change to flatbuffers 1.9
This contains two fixes for issues introduced by
IMPALA-9999's upgrade to GCC 10:
IMPALA-11599: Builds toolchain gdb with the OS compiler,
fixing an issue where it could not run on Centos 7 due
to problems with its RPATH / finding appropriate libstdc++.
IMPALA-11605: Reverts to flatbuffers 1.9, which avoids
a mismatch with the version that Hive uses. This avoids
classpath incompatibility issues.

Testing:
 - Verified that toolchain gdb works on Centos 7
 - Verified that interoperability with Hive is restored

Change-Id: Ieaafdb1da05cab60b254115faaae55dba8b67dfd
Reviewed-on: http://gerrit.cloudera.org:8080/19034
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>