IMPALA-11611: Replace THRIFT_HOME with THIFT_JAVA_HOME in fe/pom.xml
IMPALA-11369 should replace all occurrences of THRIFT_HOME with either
of THRIFT_CPP_HOME, THRIFT_JAVA_HOME, or THRIFT_PY_HOME. However, it
miss replacing THRIFT_HOME to THRIFT_JAVA_HOME in fe/pom.xml.

This reference has not cause an issue in compilation so far, but it
should be replaced with the right one. This patch does that replacement.

Testing:
- Run and pass ./buildall.sh -skiptests -notests

Change-Id: Ic39d47375353bb4d98399b5ff765d63aa331f63d
Reviewed-on: http://gerrit.cloudera.org:8080/19040
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>