IMPALA-10356: Fix SetOperationStmt::toSql() for VALUES statement with single value
If the query contains a VALUES clause with a single value, the 'analyzed
query' in the level 2 explain plan contains a UNION where both sides are
the value from the VALUES clause. For

INSERT INTO double_tbl VALUES (-0.43149576573887316);

we get

SELECT CAST(-0.43149576573887316 AS DECIMAL(17,17))
UNION
SELECT CAST(-0.43149576573887316 AS DECIMAL(17,17));

This is because of a bug in 'SetOperationStmt::toSql()'; 'ValuesStmt's
are implemented as special 'UnionStmt's, which derive from
'SetOperationStmt'. 'SetOperationStmt::toSql()' implicitly assumes that
there are at least 2 operands and always prints the first one and the
last one separately, even if they are the same.

Note that 'ValuesStmt' is the only 'SetOperationStmt' that may contain a
single operand - it is syntactically impossible in SQL to create other
'SetOperationStmt's with only one operand.

This patch modifies 'SetOperationStmt::toSql()' so that if there is only
one operand, only that operand is printed (so there is no reference to
the set operation in the resulting string). This change does not affect
how such queries are handled, only how the analyzed query is printed.

Testing:
 - Extended AnalyzeStmtsTest.TestValuesStmt so it now also checks what
   VALUES clauses with one and two operands are rewritten to during
   analysis.

Change-Id: I952377ed14eba26e3774e7776eb81a95d1d8e76f
Reviewed-on: http://gerrit.cloudera.org:8080/19048
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Daniel Becker <daniel.becker@cloudera.com>