IMPALA-14066 (Part 3): Re-applying IMPALA-11640/IMPALA-11641: Workaround errors in shared library build on Ubuntu 18+
This commit re-applies IMPALA-11640/IMPALA-11641 to the Kudu files after
the Kudu rebase to v1.17.1. The original commit message is below:

Building with -so on Ubuntu 18 or higher fails due to
an issue finding dlopen in unwind_safeness.cc:

unwind_safeness.cc:76] Check failed: !error failed to find symbol dlopen

unwind_safeness.cc is using dlsym to load the dlopen symbol
so that it can wrap it with its own dlopen code. The Impala
build has issues with the ordering of libraries, and this
code does not find dlopen. This has previously happened with
the dl_iterate_phdr symbol in Kudu.

This is a problem starting with Ubuntu 18.04, because Ubuntu 16.04
uses a version of glibc that has a bug in reporting this
error. Ubuntu 18.04 uses a newer glibc with a fix for the bug. See
https://sourceware.org/bugzilla/show_bug.cgi?id=19509 .
As a workaround for this issue, this tolerates not finding
dlopen/dlclose when building with shared libraries. Impala shared
libraries are not used in production, so this bypasses the issue.
This also adds extra validation to make sure the symbols are non-null.
Specifically, this adds another CHECK in dlsym_or_die to verify
that the symbol is non-null. This also adds a DCHECK to verify that
the symbol is non-null at dereference.

This also fixes an issue where Boost was always using static
libraries, even for shared library builds. This makes Boost use
shared libraries for shared library builds.

Testing:
 - The shared library build passes on Ubuntu 18 and Ubuntu 20
 - Impala can boot and run queries with shared libraries

Change-Id: I0034464a075b3add7ce591a36dab6fda334e6203
Reviewed-on: http://gerrit.cloudera.org:8080/19104
Reviewed-by: Daniel Becker <daniel.becker@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: http://gerrit.cloudera.org:8080/23009
Tested-by: Daniel Becker <daniel.becker@cloudera.com>