IMPALA-11617: Pool service should be made aware of processing cost limit
IMPALA-11604 enables the planner to compute CPU usage for certain
queries and to select suitable executor groups to run. The CPU usage is
expressed as the total amount of data to be processed per query.

This patch added the processing cost limit, which is the total amount of
data that each executor group can handle, to the pool service.

Testing:
 - Passed core run.
 - Verified that processing costs were shown on the admission and
   metrics pages of the Impala debug web server.

Change-Id: I9bd2a7284eda47a969ef91e4be19f96d2af53449
Reviewed-on: http://gerrit.cloudera.org:8080/19121
Reviewed-by: Qifan Chen <qchen@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>