IMPALA-11617: Pool service should be made aware of cpu core limit
IMPALA-11604 enables the planner to compute CPU usage for certain
queries and to select suitable executor groups to run. The CPU usage is
expressed as the CPU cores required to process a query.

This patch added the CPU core limit, which is the maximum CPU core
available per node and coordinator for each executor group, to the pool
service.

Testing:
 - Passed core run.
 - Verified that CPU cores were shown on the admission and
   metrics pages of the Impala debug web server.

Change-Id: Id4c5ee519ce7c329b06ac821283e215a3560f525
Reviewed-on: http://gerrit.cloudera.org:8080/19366
Reviewed-by: Andrew Sherman <asherman@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>